[FIX] Enemy speed and health buttons.

+++ src/editor_env_enemy_buttons.c
@@ -6,7 +6,7 @@
 /*   By: sipatry <sipatry@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/01/22 14:59:25 by sipatry           #+#    #+#             */
-/*   Updated: 2020/02/04 17:20:35 by sipatry          ###   ########.fr       */
+/*   Updated: 2020/02/13 12:17:41 by lnicosia         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -49,7 +49,7 @@ void	init_enemy_general_env(t_env *env)
 	env->editor.hud.g_enemy.t_health.type = INT;
 	env->editor.hud.g_enemy.t_speed.env = env;
 	env->editor.hud.g_enemy.t_speed.pos = new_point(0, 0);
-	env->editor.hud.g_enemy.t_speed.type = DOUBLE;
+	env->editor.hud.g_enemy.t_speed.type = INT;
 	env->editor.hud.g_enemy.t_scale.env = env;
 	env->editor.hud.g_enemy.t_scale.pos = new_point(0, 0);
 	env->editor.hud.g_enemy.t_scale.type = DOUBLE;
